Move parent_key concern out of logger and refactor HashFormatter
The "message" key was added in the logger, when the logger didn't have context of the format of the message. As it's more of a concern of the format, I moved it into HashFormatter. Additionally, refactored severity formats (`format_severity`) into it's own method that behaves like `format_datetime` from `::Logger::Formatter`.